requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.3.0
st-gsheets-connection==0.0.4
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
git+https://github.com/streamlit/gsheets-connection
st-gsheets-connection
datetime